clean up Command related tests
